+++ apps/web/src/config/experimentalFeatures.ts
@@ -2,6 +2,7 @@ export enum EXPERIMENTAL_FEATURES {
   WebNotifications = 'web-notifications',
   SpeedQuote = 'routing-speed-quote',
   UniversalRouter = 'universal-router',
+  PriceAPI = 'price-api',
 }
 export type EnumValues<T> = T extends { [key: string]: infer U } ? U : never
 
@@ -21,17 +22,22 @@ export type ExperimentalFeatureConfigs = FeatureRollOutConfig[]
 export const EXPERIMENTAL_FEATURE_CONFIGS: ExperimentalFeatureConfigs = [
   {
     feature: EXPERIMENTAL_FEATURES.WebNotifications,
-    percentage: 1.0,
+    percentage: 1,
     whitelist: [],
   },
   {
     feature: EXPERIMENTAL_FEATURES.SpeedQuote,
-    percentage: 0.3,
+    percentage: 0.5,
     whitelist: [],
   },
   {
     feature: EXPERIMENTAL_FEATURES.UniversalRouter,
-    percentage: 0.2,
+    percentage: 0.5,
+    whitelist: [],
+  },
+  {
+    feature: EXPERIMENTAL_FEATURES.PriceAPI,
+    percentage: 0.05,
     whitelist: [],
   },
 ]