+++ packages/uikit/src/hooks/useTooltip/types.ts
@@ -1,4 +1,4 @@
-import { Placement, Padding, PositioningStrategy } from "@popperjs/core";
+import { Padding, Placement, PositioningStrategy } from "@popperjs/core";
 
 export interface TooltipRefs {
   targetRef: React.Dispatch<React.SetStateAction<HTMLElement | null>>;