+++ packages/uikit/src/hooks/useTooltip/useTooltip.tsx
@@ -1,10 +1,10 @@
-import { AnimatePresence, Variants, LazyMotion, domAnimation } from "framer-motion";
+import { AnimatePresence, LazyMotion, Variants, domAnimation } from "framer-motion";
+import debounce from "lodash/debounce";
 import React, { useCallback, useEffect, useState } from "react";
+import { isMobile } from "react-device-detect";
 import { createPortal } from "react-dom";
 import { usePopper } from "react-popper";
-import { isMobile } from "react-device-detect";
 import { useTheme } from "styled-components";
-import debounce from "lodash/debounce";
 import getPortalRoot from "../../util/getPortalRoot";
 import isTouchDevice from "../../util/isTouchDevice";
 import { Arrow, StyledTooltip } from "./StyledTooltip";