+++ packages/utils/formatFractions.ts
@@ -1,17 +1,21 @@
-import { Percent, Fraction, Price, Currency, CurrencyAmount } from '@pancakeswap/swap-sdk-core'
+import { Currency, CurrencyAmount, Fraction, Percent, Price, Rounding } from '@pancakeswap/swap-sdk-core'
 
 export function formatPercent(percent?: Percent, precision?: number) {
   return percent ? formatFraction(percent.asFraction.multiply(100), precision) : undefined
 }
 
-export function formatFraction(fraction?: Fraction | null | undefined, precision: number | undefined = 6) {
+export function formatFraction(
+  fraction?: Fraction | null | undefined,
+  precision: number | undefined = 6,
+  rounding: Rounding | undefined = undefined,
+) {
   if (!fraction || fraction.denominator === 0n) {
     return undefined
   }
   if (fraction.greaterThan(10n ** BigInt(precision))) {
     return fraction.toFixed(0)
   }
-  return fraction.toSignificant(precision)
+  return fraction.toSignificant(precision, undefined, rounding)
 }
 
 export function formatPrice(price?: Price<Currency, Currency> | null | undefined, precision?: number | undefined) {
@@ -25,7 +29,11 @@ export function formatAmount(amount?: CurrencyAmount<Currency> | null | undefine
   if (!amount) {
     return undefined
   }
-  return formatFraction(amount?.asFraction.divide(10n ** BigInt(amount?.currency.decimals)), precision)
+  return formatFraction(
+    amount?.asFraction.divide(10n ** BigInt(amount?.currency.decimals)),
+    precision,
+    Rounding.ROUND_DOWN,
+  )
 }
 
 export function parseNumberToFraction(num: number, precision = 6) {