+++ packages/widgets-internal/farm/types.ts
@@ -1,6 +1,7 @@
 import { Token } from "@pancakeswap/sdk";
 import BigNumber from "bignumber.js";
 import { ReactNode } from "react";
+import { Address } from "viem";
 
 export interface FarmTableEarnedProps {
   earnings: number;
@@ -35,7 +36,8 @@ export interface FarmTableFarmTokenInfoProps {
   isCommunity?: boolean;
   merklLink?: string;
   hasBothFarmAndMerkl?: boolean;
-  lpAddress?: string;
+  lpAddress?: Address;
+  merklApr?: number;
 }
 
 export type ColumnsDefTypes = {